Fixed farmos planting checkbox bug
Planting checkboxes are now clickable. Also radio options are displayed if it's a single selection type.

There used to be a checkbox for the field which selects all plantings of the same field.
However, de-selecting the field's planting through the checkbox did not work (The checkbox could be unchecked, but the field's plantings would remain selected).
I am not sure how to implement this now as there are some transformations done with the fields and plantings...

So the field checkboxes are removed for now (One can still click the field to select all plantings though)