Create authorization rules for scripts. Add supertest and setup the first end to end tests for the surveystack server.
To delete a script, you must be a super admin.

To create a script, you must specify a group for the script and you must be an admin of the group that is specified.

To edit a script, you must be an admin of the group that the script is in.

To reassign the script to another group (a subset of editing), you must be an admin of the group the script is currently in and the group that the script is being reassigned to.

In the frontend, the edit script page will only display groups you are an admin of as options for what group the script can go in.

This commit sets up the first end to end tests in the surveystack server. End to end here means starting with an http request, and ending with an http response, going through all middleware, controllers, and the database (an in memory mongodb server) along the way. Assertions are made on the http response and on changes in the database. See scriptController.spec.js for the actual code.