Update idb code to be testable, awaitable, error handleable. Fix upgrade logic to not produce errors by only running things on certain database versions. Add handling of blocked and blocking scenarios so that multiple browser tabs dont interfere with eachother in the event of a database version upgrade.
Some tests in File.spec.js started failing due to issues with structuredClone. They used to pass because the idb code would fail immediately and the error was swallowed, ever getting to the actual idb operation that raises the structuredClone error. To workaround the failing tests, the store/db.js file was mocked in the File.spec.js tests. To resolve the structuredClone error we need to upgrade to jest 28, fake-indexeddb 4, and node 17. We should wait for the vue 3 branch to address that, as it will upgrade node and jest at least. Once those dependencies are upgraded, we should be able to remove the db mock from File.spec.js and the idb code thats exercised as part of the tests should run without error.