Add test coverage for createSubmission controller. Add test todos for createSubmission and updateSubmission controllers.
In the process of adding tests for creatSubmission, a bug was found wherein surveys with meta.submissions set to "group" would allow submissions to any group as long as the requesting user had access to the survey group. That bug is fixed in this commit.

I am in the process of replacing the hasEntityRights and hasSubmissionRights assertion helpers. hasEntityRights will be replaced by hasRightToManageSubmission (not commited to the codebase yet), and hasSubmissionRights will be replaced by hasRightToSubmitToSurvey (added in this commit).

This commit switches the createSubmission controller to use hasRightToSubmitToSurvey instead of hasSubmissionRights.

Subsequent commits will be adding test coverage to endpoints that utilize the helpers to to be replaced, so that we can then replace the helpers with some protection against regressions.